fix: add sorting by closed_at to analysis_engine.get_issues_close_time and analysis_engine.get_pulls_close_time
Pulls and issues are sorted by id, which is 1:1 with the created_at timestamp (although I have made the parameter explicit, just in case). This means that pulls and issues are not inherently sorted by closed_at which is what the trimming functions assume. This works for commits, because they only have one timestamp and they are sorted bby this timestamp. If a time range is provided, and so data must be trimmed, the functions now sort by closed_at in descending order before passing it to the trimming funcions. I have also added clarity to the trimming functions that they work under the assumption that the input is sorted in descending order by timestamp