feature: restructure report creation
Reports are now created in backend/analysis_engine.py and backend/crud.create_db_report now calls this, creates a models.Reports object and adds it to postgres along with the repository. backend/routers/reports.create_report has been restructured to reflect this change, and I've put the last bit in an else block for readability.