feature: change testing architecture
Tests now run from within a container that is a simple clone of the whole project. This is needed to provide reproducible results no matter which machine tests are run on - GitHub Runners get a little fuzzy when it comes to localhost and ports and such, this should alleviate this entirely as tests are now done exclusively in the docker network. Also, non-private environment variables like POSTGRES_USER have been removed from .env and are just dropped in directly into docker compose. This only applies to a dev or test context though - for prod, these values will be funnelled in through GitHub Secrets.