fix: add routing via nginx reverse proxy
Now when an api request is made to the backend container, it will route appropriately depending on the application context. In prod, it will route the request into the backend container via the docker network. In dev, it will route the request by calling localhost, since the containers all have exposed ports in development. Maybe this needs to be changed in the future?