feature: rewrite code churn function to use the new get_contributor_history function
This is a truer code churn value because it omits merge commits and empty commits. It is extremely rate limit efficient, with the only caveat being it is limited to 10,000 commits per repository. This is actually double the limit of the old function, which would've wiped out the rate limit should it have ever encountered a repository that big. This new function will make at most 7 API requests and GitHub caches this result, so in some cases only 2 API requests are made. This is regardless of how many commits the target repository has, and the speed of this function is extremely quick compared to the old function.