feature: change how code churn validates repo size
I forgot that checking the commit count of a repository is actually quite rate limit expensive (100 commits = 1 API call), so my previous claim of code churn not exceeding 7 API calls was wrong. Now, however, the commit size is checked from the /stats/contributors response directly. The endpoint returns commit count regardless of the repository's size, so the code churn function tallies it up and checks afterwards if it has exceeded 10,000. If it has, that means that the repository was too big, and the code churn function returns a success: False key-value alongside the code churn. This means that if the code churn is 0, we know if it is because there are no commits, or if it is because there are too many commits, and the frontend can differentiate and provide the appropriate response.