feature: rework report parsing in the frontend
CreateReport.jsx now uses the new success value provided in code churn to inform the user if their repository is too big for code churn to work. This allows the other metrics to still be calculated and displayed even if code churn needs to be omitted. I've tested this on a repository with 20,000 commits and it worked perfectly.