feature: catch errors and parse them into the report, restore previous code churn return structure
Instead of errors being raised and crashing report creation, they are caught and parsed as strings inside analysis_engine.create_report, and are returned as part of the report in place of the corresponding value. Restored the original format of the get_code_churn function response, as the error catching is more intuitive, and changing the format of data gets messy in the database. Reports can now be generated partially, in cases where a repository may for example not have issues or pull requests yet.