feature: implement upsert for concurrent identical requests
Now when two or more report requests are sent for the same repository in a short span of time, the backend can resolve the conflict appropriately, so duplicate rows aren't created. A semaphore was also implemented to queue 4 concurrent API requests at a time, which stops the GitHub API terminating the connection when it thinks it's getting a DoS attack. This introduces complications for concurrent report generations, as it effectively makes it synchronous (the individual metrics are still calculated asynchronously). This will be explained more in the pull request.