feature: remove time filtering from backend/analysis_engine.py
This was the workaround to the non-granular time filtering that the REST API had to offer, but now with the plans of switching to GitHub's GraphQL API, the backend will have access to much more granular time filtering directly in the request, so manually filtering data won't be necessary. Time filtering doesn't have any use yet, but it's intended to allow users to generate time filtered reports.